Don't create shadow copy that already exists (#147)
In some scenarios the same unmanaged DLL is resolved multiple times
(e.g. when using System.Data.SQLite.SQLiteConnection). The previous
implementation would then try to create new shadow copies of the same
DLL during subsequent resolves. This failed during copy of the DLL since
the parameter to enable overwriting was not set to true.

This fix checks if the shadow copy already exists and skips the copy
step if it does. The path to the existing shadow copy is returned
instead.

Fixes #146

+++ src/Plugins/Loader/ManagedLoadContext.cs
@@ -343,7 +343,10 @@ namespace McMaster.NETCore.Plugins.Loader
             var dllFileName = Path.GetFileName(dllPath);
             var shadowCopyPath = Path.Combine(_unmanagedDllShadowCopyDirectoryPath, dllFileName);
 
-            File.Copy(dllPath, shadowCopyPath);
+            if (!File.Exists(shadowCopyPath))
+            {
+                File.Copy(dllPath, shadowCopyPath);
+            }
 
             return shadowCopyPath;
         }
